requirements.txt
amqp==5.3.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
asttokens==3.0.1
billiard==4.2.4
boto3==1.42.41
botocore==1.42.41
brotli==1.2.0
celery==5.6.2
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.13.3
crispy-bootstrap5==2025.6
cron_descriptor==2.0.6
cssselect2==0.8.0
decorator==5.2.1
diff-match-patch==20241021
Django==5.2.11
django-allauth==65.14.0
django-axes==8.1.0
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-environ==0.12.0
django-filter==25.2
django-import-export==4.4.0
django-simple-history==3.11.0
django-storages==1.14.6
django-timezone-field==7.2.1
django-widget-tweaks==1.5.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
executing==2.2.1
factory_boy==3.3.3
Faker==40.1.2
fonttools==4.57.0
gunicorn==25.0.1
html5lib==1.1
iniconfig==2.3.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.1.0
kombu==5.6.2
matplotlib-inline==0.2.1
numpy==2.4.2
packaging==26.0
pandas==3.0.0
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg==3.3.2
psycopg-binary==3.3.2
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
pydyf==0.6.0
Pygments==2.19.2
PyJWT==2.11.0
pypdf==6.6.2
pyphen==0.17.2
pytest==9.0.2
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
redis==7.1.0
reportlab==4.4.10
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
tablib==3.9.0
tinycss2==1.4.0
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.5.3
weasyprint==59.0
webencodings==0.5.1
whitenoise==6.11.0
zopfli==0.4.0
openpyxl==3.1.5
